RPCT-3203, u7, lib: PropertyEditor is compatible with both Qt 5 and Qt 6, minor type casts added

+++ UtilsLib/LogFile.cpp
@@ -1105,7 +1105,7 @@ namespace Log
 
 		if (role == Qt::DisplayRole)
 		{
-			int displayIndex = column;
+			int displayIndex = static_cast<int>(column);
 			if (displayIndex == m_columnTime)
 			{
 				return record.time.toString(messageTimeFormat);
